move test_qweb_float from test_orm to test_base
