+++ odoo/addons/test_base/security/ir.model.access.csv
@@ -9,3 +9,4 @@ access_emailmessage,test_orm_emailmessage,test_base.model_test_orm_emailmessage,
 access_mixed,test_orm_mixed,test_base.model_test_orm_mixed,base.group_user,1,1,1,1
 access_binary_test,access_binary_test,model_binary_test,base.group_user,1,0,0,0
 access_calendar_test,access_calendar_test,model_calendar_test,base.group_user,1,0,0,0
+access_decimal_precision_test_all,decimal.precision.test,model_decimal_precision_test,base.group_system,1,1,1,1
