+++ odoo/addons/test_orm/security/ir.model.access.csv
@@ -1,6 +1,5 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
 access_category,test_orm_category,test_orm.model_test_orm_category,base.group_user,1,1,1,1
-access_decimal_precision_test_all,decimal.precision.test,model_decimal_precision_test,base.group_system,1,1,1,1
 access_discussion,test_orm_discussion,test_orm.model_test_orm_discussion,base.group_user,1,1,1,1
 access_message,test_orm_message,test_orm.model_test_orm_message,base.group_user,1,1,1,1
 access_emailmessage,test_orm_emailmessage,test_orm.model_test_orm_emailmessage,base.group_system,1,1,1,1
